refactor : base.html 위치 변경, package 변경
- base.html 위치 변경
다른 경로에서도 사용할 수 있도록 root 로 위치 변경

- 모호한 package 명 변경
web -> user
