+++ src/main/java/com/codessquad/qna/user/User.java
@@ -1,4 +1,4 @@
-package com.codessquad.qna.web;
+package com.codessquad.qna.user;
 
 public class User {
   private String userId;
