+++ src/main/java/com/codessquad/qna/user/UserController.java
@@ -1,4 +1,4 @@
-package com.codessquad.qna.web;
+package com.codessquad.qna.user;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
