+++ src/main/resources/templates/user/form.html
@@ -27,4 +27,4 @@
 </div>
 {{/partial}}
 
-{{> user/base}}
+{{> base}}
