+++ src/main/resources/templates/user/list.html
@@ -30,4 +30,4 @@
 </div>
 {{/partial}}
 
-{{> user/base}}
+{{> base}}
