+++ src/main/resources/templates/user/profile.html
@@ -26,4 +26,4 @@
 </div>
 {{/partial}}
 
-{{> user/base}}
+{{> base}}
